requirements.txt
absl-py==1.4.0
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.2
googlemaps==4.10.0
h11==0.14.0
idna==3.4
numpy==1.24.3
ortools==9.6.2534
protobuf==4.23.1
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.30.0
scipy==1.10.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
